Move ExecutionEngineService.noOpApi to tests package.
