+++ casper/src/test/scala/io/casperlabs/casper/helper/HashSetCasperTestNode.scala
@@ -291,7 +291,7 @@ object HashSetCasperTestNode {
                 _ =>
                   n.connectionsCell.flatModify(
                     _.addConn[F](m.local)(Monad[F], n.logEff, n.metricEff)
-                )
+                  )
               )
           }
     } yield nodes
